scripts/inference/requirements.txt
numpy
pandas
tqdm
pyyaml
